Fix reading literals when last line is in next ByteBuf.
Summary of the current bug:
==============================
Existing code assumes the last line will come in the same bytebuf as the literals.  It could be the next bytebuf, so we have to process next bytebuf when this happens.

Due to existing code assumption, it was throwing NPE in line 89 (because there was no more bytes to read),  even NPE is avoided, it still misses consuming ")\r\n" in the next bytebuf because there is no code to handle that.